Refactor: Rename 경로 재수정
