+++ __tests__/ValidateTest.js
@@ -1,5 +1,4 @@
 const MissionUtils = require("@woowacourse/mission-utils");
-const App = require("../src/App");
 const {
   bridgeLength,
   bridgeDirection,
